Remove some tests moved to golem
