+++ test/unit/dag/test_graph_operator.py
@@ -1,12 +1,6 @@
-from copy import deepcopy
-
-from golem.core.dag.graph_utils import distance_to_root_level, nodes_from_layer
-from golem.core.dag.linked_graph import LinkedGraph, get_distance_between
-from golem.core.optimisers.graph import OptNode
-
-from fedot.core.pipelines.adapters import PipelineAdapter
 from fedot.core.pipelines.node import PrimaryNode, SecondaryNode
 from fedot.core.pipelines.pipeline import Pipeline
+from golem.core.optimisers.graph import OptNode
 
 
 def get_pipeline() -> Pipeline:
